# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"MissionResponse": {
				"description": "미션 정보",
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "미션 ID"
					},
					"storeId": {
						"type": "number",
						"format": "double",
						"description": "미션 등록 가게 ID"
					},
					"title": {
						"type": "string",
						"description": "미션 제목"
					},
					"reward": {
						"type": "number",
						"format": "double",
						"description": "미션 보상 포인트"
					},
					"deadline": {
						"type": "string",
						"description": "미션 마감일"
					},
					"createdAt": {
						"type": "string",
						"description": "미션 생성 시각"
					}
				},
				"required": [
					"id",
					"storeId",
					"title",
					"reward",
					"deadline",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiSuccessResponse_MissionResponse_": {
				"description": "API 성공 응답 공통 포맷",
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false,
						"description": "요청 처리 결과"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true,
						"description": "성공 응답 null 값"
					},
					"data": {
						"$ref": "#/components/schemas/MissionResponse",
						"description": "API별 성공 응답 데이터"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiErrorData": {
				"description": "에러 응답 추가 데이터",
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"anyOf": [
						{
							"type": "string"
						},
						{
							"type": "number",
							"format": "double"
						},
						{
							"type": "boolean"
						}
					],
					"nullable": true
				}
			},
			"ApiErrorResponseBody": {
				"description": "API 실패 응답 에러 상세 정보",
				"properties": {
					"errorCode": {
						"type": "string",
						"description": "에러 코드"
					},
					"message": {
						"type": "string",
						"description": "클라이언트 전달 에러 메시지"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/ApiErrorData"
							}
						],
						"nullable": true,
						"description": "에러 관련 추가 데이터"
					}
				},
				"required": [
					"errorCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiFailResponse": {
				"description": "API 실패 응답 공통 포맷",
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"FAIL"
						],
						"nullable": false,
						"description": "요청 처리 결과"
					},
					"error": {
						"$ref": "#/components/schemas/ApiErrorResponseBody",
						"description": "실패 원인 상세 정보"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true,
						"description": "실패 응답 null 값"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateMissionRequest": {
				"description": "미션 생성 요청",
				"properties": {
					"title": {
						"type": "string",
						"description": "미션 제목"
					},
					"reward": {
						"type": "number",
						"format": "double",
						"description": "미션 보상 포인트"
					},
					"deadline": {
						"type": "string",
						"description": "미션 마감일"
					}
				},
				"required": [
					"title",
					"reward",
					"deadline"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiSuccessResponse_MissionResponse-Array_": {
				"description": "API 성공 응답 공통 포맷",
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false,
						"description": "요청 처리 결과"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true,
						"description": "성공 응답 null 값"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/MissionResponse"
						},
						"type": "array",
						"description": "API별 성공 응답 데이터"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewResponse": {
				"description": "리뷰 정보",
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "리뷰 ID"
					},
					"storeId": {
						"type": "number",
						"format": "double",
						"description": "리뷰 등록 가게 ID"
					},
					"memberId": {
						"type": "number",
						"format": "double",
						"description": "리뷰 작성 회원 ID"
					},
					"rating": {
						"type": "number",
						"format": "double",
						"description": "평점"
					},
					"content": {
						"type": "string",
						"description": "리뷰 내용"
					},
					"createdAt": {
						"type": "string",
						"description": "리뷰 생성 시각"
					}
				},
				"required": [
					"id",
					"storeId",
					"memberId",
					"rating",
					"content",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiSuccessResponse_ReviewResponse_": {
				"description": "API 성공 응답 공통 포맷",
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false,
						"description": "요청 처리 결과"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true,
						"description": "성공 응답 null 값"
					},
					"data": {
						"$ref": "#/components/schemas/ReviewResponse",
						"description": "API별 성공 응답 데이터"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateReviewRequest": {
				"description": "리뷰 생성 요청",
				"properties": {
					"rating": {
						"type": "number",
						"format": "double",
						"description": "평점, 1부터 5까지"
					},
					"content": {
						"type": "string",
						"description": "리뷰 내용"
					}
				},
				"required": [
					"rating",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StoreResponse": {
				"description": "가게 정보",
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "가게 ID"
					},
					"regionId": {
						"type": "number",
						"format": "double",
						"description": "가게 소속 지역 ID"
					},
					"name": {
						"type": "string",
						"description": "가게 이름"
					},
					"address": {
						"type": "string",
						"description": "가게 주소"
					}
				},
				"required": [
					"id",
					"regionId",
					"name",
					"address"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiSuccessResponse_StoreResponse_": {
				"description": "API 성공 응답 공통 포맷",
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false,
						"description": "요청 처리 결과"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true,
						"description": "성공 응답 null 값"
					},
					"data": {
						"$ref": "#/components/schemas/StoreResponse",
						"description": "API별 성공 응답 데이터"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateStoreRequest": {
				"description": "가게 생성 요청",
				"properties": {
					"name": {
						"type": "string",
						"description": "가게 이름"
					},
					"address": {
						"type": "string",
						"description": "가게 주소"
					}
				},
				"required": [
					"name",
					"address"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "umc-node-simple-api",
		"version": "1.0.0",
		"description": "Simple Express API server with SQLite dummy data.",
		"contact": {}
	},
	"paths": {
		"/stores/{storeId}/missions": {
			"post": {
				"operationId": "CreateMission",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_MissionResponse_"
								}
							}
						}
					},
					"400": {
						"description": "storeId, title, reward, deadline이 유효하지 않은 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					},
					"401": {
						"description": "로그인이 필요하거나 토큰이 유효하지 않은 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 가게인 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					}
				},
				"description": "특정 가게 미션 등록",
				"tags": [
					"Stores"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "가게 ID",
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "미션 생성 요청 body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateMissionRequest",
								"description": "미션 생성 요청 body"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetMissions",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_MissionResponse-Array_"
								}
							}
						}
					},
					"400": {
						"description": "storeId, limit, cursor가 유효하지 않은 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 가게인 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					}
				},
				"description": "특정 가게 등록 미션 목록 조회",
				"tags": [
					"Stores"
				],
				"security": [],
				"parameters": [
					{
						"description": "가게 ID",
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "한 번에 조회할 개수, 기본값 10, 최대 50",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "이전 페이지 마지막 미션 ID",
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/stores/{storeId}/reviews": {
			"post": {
				"operationId": "CreateReview",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_ReviewResponse_"
								}
							}
						}
					},
					"400": {
						"description": "storeId, rating, content가 유효하지 않은 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					},
					"401": {
						"description": "로그인이 필요하거나 토큰이 유효하지 않은 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 가게인 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					}
				},
				"description": "특정 가게 리뷰 작성",
				"tags": [
					"Stores"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "가게 ID",
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "리뷰 생성 요청 body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateReviewRequest",
								"description": "리뷰 생성 요청 body"
							}
						}
					}
				}
			}
		},
		"/regions/{regionId}/stores": {
			"post": {
				"operationId": "CreateStore",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_StoreResponse_"
								}
							}
						}
					},
					"400": {
						"description": "regionId, name, address가 유효하지 않은 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 지역인 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					}
				},
				"description": "특정 지역 가게 등록",
				"tags": [
					"Stores"
				],
				"security": [],
				"parameters": [
					{
						"description": "지역 ID",
						"in": "path",
						"name": "regionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "가게 생성 요청 body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateStoreRequest",
								"description": "가게 생성 요청 body"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}